added validation for inactive user

+++ src/controllers/user.controller.js
@@ -82,6 +82,8 @@ const loginUser = async (req, res) => {
       return res
         .status(400)
         .json({ message: "La contraseña ingresada no es valida" });
+      
+    if(!user.active) return res.status(400).json({ message: "El usuario ha sido deshabilitado por un administrador" });
 
     sendToken(user, 201, res);
   } catch (error) {
@@ -140,12 +142,16 @@ const getAllUsers = async (req, res) => {
 const editUser = async (req, res) => {
   try {
     const { id } = req.params;
+
+    if(id == req.user.id) return res.status(400).json({ message: "No puedes editar tu propio usuario" });
+
     const payload = req.body;
+
     const userEdited = await editUserService(id, payload);
     if (!userEdited) return res.status(404).json("Usuario no encontrado");
     res
       .status(200)
-      .json({ message: "Usuario editado con exito", user: userEdited });
+      .json(userEdited);
   } catch (error) {
     res.status(500).json({ message: "Error al editar el usuario", error });
   }
@@ -155,6 +161,8 @@ const deleteUser = async (req, res) => {
   try {
     const { id } = req.params;
 
+    if(id == req.user.id) return res.status(400).json({ message: "No puedes eliminar tu propio usuario" });
+
     const userDeleted = await deleteUserService(id);
 
     if (!userDeleted) {
@@ -163,7 +171,7 @@ const deleteUser = async (req, res) => {
 
     res
       .status(200)
-      .json({ message: "Usuario eliminado con exito", user: userDeleted });
+      .json(userDeleted);
   } catch (error) {
     res.status(500).json({ message: "Error al eliminar el usuario", error });
   }
